extracted webdriver creation to a separate module
