+++ executor-service/src/main/resources/application.properties
@@ -1,8 +1,4 @@
-executorservice.common.driverWait=3
-executorservice.common.pageLoadTimeout=30000
 executorservice.common.threadsCount=1
-executorservice.common.webDriverExecutable=/opt/chromedriver
-executorservice.common.userAgent=desktop
 executorservice.common.keepAliveTime=30000
 publisher.proxy.url=https://publisher-microservice-a94942799a97.herokuapp.com/publisher/proxy/count/5
 publisher.scenario.url=https://publisher-microservice-a94942799a97.herokuapp.com/publisher/scenario/count/5
