+++ settings.gradle.kts
@@ -3,3 +3,4 @@ include("executor-service")
 include("logback")
 include("logger")
 include("model")
+include("webdriver")
