+++ webdriver/src/main/java/executor/service/webdriver/factory/proxy/ProxyProvider.java
@@ -1,4 +1,4 @@
-package executor.service.factory.webdriverinitializer.proxy;
+package executor.service.webdriver.factory.proxy;
 
 import executor.service.model.ProxyConfigHolder;
 import org.openqa.selenium.Proxy;
