+++ webdriver/src/main/java/executor/service/webdriver/factory/setting/BrowserOptions.java
@@ -1,4 +1,4 @@
-package executor.service.factory.webdriverinitializer.setting;
+package executor.service.webdriver.factory.setting;
 
 public enum BrowserOptions {
     DISABLE_SANDBOX("--no-sandbox"),
