+++ webdriver/src/main/resources/webdriver.properties
@@ -1,0 +1,4 @@
+executorservice.common.driverWait=3
+executorservice.common.pageLoadTimeout=30000
+executorservice.common.webDriverExecutable=/opt/chromedriver
+executorservice.common.userAgent=desktop
